Fixed `parser` example's type
The return value of `input` must match the stored `Producer`'s value, so
when you compose `input` with `take` it infers a `()` return value for `input`,
which translates to a `()` return value for the producer.

The alternative fix would be to wrap the `input` in `void` but I figured that
changing the type was simpler and required less explanation.